perf(export,processing): drop a dead entity index, share the one already built
`relationships()` built a full parallel entity index and handed it to the
decoder, then never read it. Every decode in that function is
`decode_at_with_id` over spans the scanner already produced; only
`decode_by_id` consults an index, and nothing there calls it. The build was a
whole extra scan of the file for a map nobody looked at.

`extract_georeferencing` does need its index, because the extractor resolves
references by id. What it did not need was to build its own when the caller had
one. It gains `extract_georeferencing_with_index`, and `process_geometry` now
passes the index it has held since the start of the run. The old signature stays
as a thin wrapper, so no caller changes.

Measured with the harness this adds, best of 3, on release fixtures (ms):

| fixture                | parallel index | bare type scan | relationships() before | after |
|------------------------|---------------:|---------------:|-----------------------:|------:|
| O-S1-BWK (327 MB)      |           59.7 |          190.7 |                  234.6 | 181.1 |
| ISSUE_053 (169 MB)     |           35.6 |           88.5 |                  118.9 |  86.4 |
| ISSUE_068 (54 MB)      |           11.7 |           29.0 |                   40.8 |  28.2 |
| schependomlaan (47 MB) |            8.6 |           28.4 |                   38.1 |  27.1 |
| advanced_model (34 MB) |            6.6 |           16.5 |                   22.3 |  16.3 |

`relationships()` now lands just under the bare scan on every fixture, which is
what it should cost. The georeferencing half saves one index build per
`process_geometry` run, about 57 ms on the largest fixture.

Size it honestly: this is roughly 1% of a large conversion. The scan is the
bigger half of both helpers and this does not touch it. The reason to take it is
that it is free and it applies to every model.

Output is unchanged by construction. The removed index was never read. The
shared one is the same map the callee built for itself, and `parallel ==
serial` is already pinned by the parallel_scan tests. Two new tests hold the
georeferencing half: one pins the variant equal to the wrapper, one passes an
empty index and asserts the result collapses, so reintroducing an internal build
fails rather than going unnoticed.

No test can honestly guard the `relationships()` deletion. A dead index has no
output effect, timing assertions are flake, and source-text assertions are out.
The comment in that function is the guard.

`probe.sh` reads flat on this diff by construction: `ProcessingStats` closes its
timers before the metadata block that calls georeferencing runs. That gap is now
recorded in the perf ledger alongside the other two, so a flat table there is
read as a control rather than a result.

`georeferencing.rs` sat at exactly its 468-line ratchet budget, so its tests move
to a sibling `georeferencing_tests.rs` following the `stream_meta.rs` pattern and
the module leaves the allowlist at 162 lines.
